Pass config via the environment
This avoids the need to authorize the token and avoids the need for a
volume for storing data between runs.

We disable SSH host key checking because it is error prone to do it
manually. The Platform.sh CLI command uses ssh certificates and we don't
send any confidential info over the SSH channel.