Copy platformsh-cli from another Docker image
No need to curl it ourselves.

This will trigger a pull request from @dependabot whenever there is a
new version of platformsh-cli (that isn't be the case with curl).

There is no tagged releases of
https://hub.docker.com/r/pjcdawkins/platformsh-cli, though. Only
`latest`.

But apparently `FROM` lines can references the digest of the image _and_
@dependabot should be able to update based on digests as well (TIL).

Although https://hub.docker.com/r/pjcdawkins/platformsh-cli doesn't
appear to an official platformsh-cli image, @pjcdawkins is an employee
of Platform.sh and the image appears to get updates fast and automated.